use floats instead of ints for sizes
git-svn-id: https://ssl.drqueue.org/svn/branches/0.64.x@2984 94d6b6b3-0ee6-0310-ab95-9b4161196ee2

+++ DrQueueOnRails/app/controllers/jobs_controller.rb
@@ -166,13 +166,13 @@ ENV['WEB_PROTO']+"://")
    		quota = 0
    		status_arr.each do |stat|
    		  if profile.status == stat
-   		    quota = quota_arr[i].to_i
+   		    quota = quota_arr[i].to_f
    		  end
    		  i += 1
    		end 
    		
    		# userdir size in KB
-    	du = `du -s #{userdir} | awk '{print $1}'`.to_i	
+    	du = `du -s #{userdir} | awk '{print $1}'`.to_f	
     	used = du / 1048576.0
     	
     	if used > quota
